requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.40.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.6
cssselect==1.2.0
Django==5.0.4
django-celery-results==2.5.1
django-cloudinary-storage==0.3.0
django-redis==5.4.0
dnspython==2.6.1
eventlet==0.36.1
filelock==3.14.0
flower==2.0.1
greenlet==3.0.3
gunicorn==22.0.0
humanize==4.9.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.2.0
jmespath==1.0.1
kombu==5.3.7
lxml==5.2.1
packaging==24.0
parsel==1.9.1
pillow==10.3.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
Protego==0.3.1
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2024.1
queuelib==1.7.0
redis==5.0.4
requests==2.31.0
requests-file==2.0.0
Scrapy==2.11.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
tldextract==5.1.2
tornado==6.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
w3lib==2.1.2
wcwidth==0.2.13
whitenoise==6.6.0
zope.interface==6.3
